requirements.txt
# Core RAG Dependencies
langchain>=0.1.0
langchain-community>=0.0.20

# Vector Store
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.3.1

# LLM Integration
google-genai>=0.2.0

# PDF Processing
pypdf>=4.0.0
pymupdf>=1.23.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorama>=0.4.6
